Uprava nekterych souboru

+++ CRUD/Database/Product.cs
@@ -9,7 +9,7 @@ namespace CRUD
         public const string PriceString = "price";
         public const string DescriptionString = "description";
 
-        public Product(long id, string name, long categoriesID, double price, string description)
+        public Product(long id, string name, long categoriesID, long price, string description)
         {
             Id = id;
             Name = name;
@@ -21,7 +21,7 @@ namespace CRUD
         public long Id { get; set; }
         public string Name { get; set; }
         public long CategoriesId { get; set; }
-        public double Price { get; set; }
+        public long Price { get; set; }
         public string Description { get; set; }
 
         public override string ToString()
